apps/osc/scripts/login_and_create_quote.py
"""Login and create quote workflow for OSC application.

Demonstrates complete automation workflow:
1. Launch browser
2. Login to OSC
3. Navigate to quote creation
4. Fill and submit quote form
5. Handle errors with screenshots and logging
"""

import json
import sys
from pathlib import Path

# Add project root to path
project_root = Path(__file__).parent.parent.parent.parent
sys.path.insert(0, str(project_root))

from dotenv import load_dotenv

from apps.osc.pages.dashboard_page import DashboardPage
from apps.osc.pages.login_page import LoginPage
from apps.osc.pages.quote_page import QuotePage
from core.browser import BrowserManager
from core.config import settings
from core.logger import Logger
from core.utils import ensure_dir, get_env, now_ts

# Load environment variables
load_dotenv()


def load_test_data() -> dict:
    """Load test data from sample inputs JSON.

    Returns:
        Dictionary containing test data
    """
    data_file = Path(__file__).parent.parent / "data" / "sample_inputs.json"

    with open(data_file) as f:
        return json.load(f)


def main() -> None:
    """Execute OSC login and quote creation workflow."""
    logger = Logger.get("osc")
    browser_manager = BrowserManager()

    logger.info("=" * 80)
    logger.info("Starting OSC automation workflow: Login and Create Quote")
    logger.info(f"Settings | headless={settings.headless}, trace={settings.trace_enabled}")
    logger.info("=" * 80)

    try:
        # Load test data
        test_data = load_test_data()
        logger.info("Test data loaded successfully")

        # Get credentials from environment or use test data defaults
        username = get_env("OSC_USER", test_data["username"])
        password = get_env("OSC_PASS", test_data["password"])

        # Launch browser and create context
        browser_manager.launch()
        context = browser_manager.new_context()
        page = browser_manager.new_page(context)

        # Initialize page objects
        login_page = LoginPage(page)
        dashboard_page = DashboardPage(page)
        quote_page = QuotePage(page)

        # Step 1: Login
        logger.info("Step 1: Performing login")
        login_page.open()
        login_page.login(username, password)
        logger.info("✓ Login completed successfully")

        # Step 2: Navigate to dashboard and verify
        logger.info("Step 2: Verifying dashboard access")
        dashboard_page.wait_for_dashboard_loaded()
        logger.info("✓ Dashboard loaded successfully")

        # Step 3: Navigate to quote creation
        logger.info("Step 3: Navigating to quote creation")
        dashboard_page.navigate_to_quotes()
        logger.info("✓ Navigation to quotes successful")

        # Step 4: Create quote (if form is accessible)
        logger.info("Step 4: Attempting to create quote")
        try:
            quote_page.create_quote(
                customer_name=test_data["customer"]["name"],
                customer_email=test_data["customer"]["email"],
                customer_phone=test_data["customer"]["phone"],
                product=test_data["product"]["value"],
                quantity=test_data["product"]["quantity"],
            )
            logger.info("✓ Quote creation completed successfully")
        except Exception as e:
            logger.warning(f"Quote creation not available (expected for demo): {e}")
            logger.info("✓ Workflow navigated to quote section (demo endpoint may not exist)")

        # Success screenshot
        screenshot_dir = ensure_dir(Path(__file__).parent.parent / "reports")
        screenshot_path = screenshot_dir / f"success_{now_ts()}.png"
        login_page.ui.screenshot(str(screenshot_path))

        logger.info("=" * 80)
        logger.info("✅ OSC workflow completed successfully")
        logger.info("=" * 80)

    except Exception as e:
        logger.error("=" * 80)
        logger.error(f"❌ Workflow failed: {e}")
        logger.error("=" * 80)

        # Error screenshot
        try:
            screenshot_dir = ensure_dir(Path(__file__).parent.parent / "reports")
            screenshot_path = screenshot_dir / f"error_{now_ts()}.png"
            login_page.ui.screenshot(str(screenshot_path))
            logger.info(f"Error screenshot saved | path={screenshot_path}")
        except Exception as screenshot_error:
            logger.warning(f"Failed to capture error screenshot: {screenshot_error}")

        raise

    finally:
        # Always cleanup browser resources
        logger.info("Cleaning up browser resources")
        browser_manager.close()
        logger.info("Browser closed")


if __name__ == "__main__":
    try:
        main()
        sys.exit(0)
    except Exception as e:
        print(f"Script failed: {e}", file=sys.stderr)
        sys.exit(1)
